% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_function.R
\name{influenceR}
\alias{influenceR}
\title{influenceR}
\usage{
influenceR(model, data, plot = "none")
}
\arguments{
\item{model}{a glm or lm object}

\item{data}{a dataframe}

\item{plot}{optional parameter to include plots in the output, default is "none", options include none", "all", "Cooks", "DFFITS", "Hadi"}
}
\value{
Returns a dataframe with all influential measures. If `plot` parameter is set to anything other than "none", the function will return a list with dataframe and `ggplot2` plot objects.
}
\description{
influenceR
}
\examples{
#import dataset mtcars
data1 <- mtcars
model1 <- lm(mpg ~ wt + hp, data = data1)
sample1 <- influenceR(model1, data1, "all")

## show dataframe of measures
sample1$influence_measures

## show plots
sample1$plots

## show individual plots
sample1$plots$Cooks
sample1$plots$DFFITS
sample1$plots$Hadi

}
